src/requirements.txt
python-dotenv
langchain
langchain-core
langchain-openai
langchain-community
tiktoken
wikipedia
streamlit
streamlit-chat
pyodbc
sqlalchemy
pandas
azure.identity